+++ askbot/doc/source/changelog.rst
@@ -4,6 +4,7 @@ Changes in Askbot
 Development version (not released yet)
 --------------------------------------
 * Made email recovery link work when askbot is deployed on subdirectory (Evgeny)
+* Added tests for the CSRF_COOKIE_DOMAIN setting in the startup_procedures (Evgeny)
 
 0.7.39 (Jan 11, 2012)
 ---------------------